% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content-uri.R
\name{content_id}
\alias{content_id}
\title{Generate a content uri for a local file}
\usage{
content_id(file, open = "", raw = TRUE)
}
\arguments{
\item{file}{path to the file, URL, or a \link[base:file]{base::file} connection}

\item{open}{The mode to open text, see details for \code{Mode} in \link[base:file]{base::file}.}

\item{raw}{Logical, should compressed data be left as compressed binary?}
}
\value{
a content identifier uri
}
\description{
Generate a content uri for a local file
}
\details{
See \url{https://github.com/hash-uri/hash-uri} for an overview of the
content uri format and comparison to similar approaches.

Compressed file streams will have different raw (binary) and uncompressed
hashes. Set \code{raw = FALSE} to allow \link[base:file]{base::file} connection to uncompress
common compression streams before calculating the hash, but this will
be slower.
}
\examples{
path <- tempfile("iris", , ".csv")
write.csv(iris, path)
content_id(path)

## Note that a different serialization gives a different hash:
path_txt <- tempfile("iris", , ".txt")
write.table(iris, path_txt)
content_id(path_txt)

}
