% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content-uri.R
\name{content_uri}
\alias{content_uri}
\title{Generate a content uri for a local file}
\usage{
content_uri(path, raw = TRUE, ...)
}
\arguments{
\item{path}{path to the file}

\item{raw}{logical, whether the content should be for the raw file
or contents, see \link[base:file]{base::file}}

\item{...}{additional arguments to \link[base:file]{base::file}}
}
\value{
a content identifier uri
}
\description{
Generate a content uri for a local file
}
\details{
See \url{https://github.com/hash-uri/hash-uri} for an overview of the
content uri format and comparison to similar approaches.

Compressed file streams will have different raw (binary) and uncompressed
hashes. Set \code{raw = FALSE} will allow \link{file} connection to uncompress
common compression streams before calculating the hash, but this will
be slower.
}
\examples{
path <- tempfile("iris", , ".csv")
write.csv(iris, path)
content_uri(path)

## Note that a different serialization gives a different hash:
path_txt <- tempfile("iris", , ".txt")
write.table(iris, path_txt)
content_uri(path_txt)
}
