% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdb_registry.R
\name{default_lmdb}
\alias{default_lmdb}
\title{default location for LMDB registry}
\usage{
default_lmdb(dir = content_dir())
}
\arguments{
\item{dir}{base directory for LMDB}
}
\description{
Helper utility to initialize an LMDB registry -
matcher uses the pattern: "any file path ending in "lmdb".
The default map size can be set using, e.g.
\code{options(thor_mapsize=1e12)}
}
\details{
Windows machines may need to set a smaller map size,
see \code{thor::mdb_env} for details.
}
