% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_registries.R
\name{default_registries}
\alias{default_registries}
\title{default registries}
\usage{
default_registries()
}
\description{
A helper function to conveniently load the default registries
}
\details{
This function is primarily useful to restrict the
scope of \link{sources} or \link{register} to, e.g. either just the
remote registry or just the local registry.  Note that a user
can alter the registry on the fly by passing local paths and/or the
URL (\verb{https://hash-archive.org}) directly.
}
\examples{
## Both defaults
default_registries()

## Only the fist one (local registry)
default_registries()[1]
\donttest{
## Alter the defaults with env var.
Sys.setenv(CONTENTID_REGISTRIES = tempfile())
default_registries()

Sys.unsetenv("CONTENTID_REGISTRIES")
}
}
