% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_resource.R
\name{has_resource}
\alias{has_resource}
\title{has_resource}
\usage{
has_resource(url = NULL)
}
\arguments{
\item{url}{vector of URL resources required}
}
\description{
Helper function to ensure examples do not execute when internet
resource is temporarily unavailable, as in such cases rendering
the example does not provide a reliable check.  This allows
examples ("tests") to "fail gracefully".
}
\examples{
has_resource("https://google.com")

}
