% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash-uri.R
\name{hash_uri}
\alias{hash_uri}
\title{Generate a hash uri for a local file}
\usage{
hash_uri(path, raw = TRUE, ...)
}
\arguments{
\item{path}{path to the file}

\item{raw}{logical, whether the hash should be for the raw file or contents, see [base::file]}

\item{...}{additional arguments to [base::file]}
}
\value{
a hash uri identifier (which can be resolved by [request_url])
}
\description{
Generate a hash uri for a local file
}
\details{
See <https://github.com/hash-uri/hash-uri> for an overview of the hash uri format.

Compressed file streams will have different raw (binary) and uncompressed hashes.
Set `raw = FALSE` will allow [file] connection to uncompress common
compression streams before calculating the hash, but this will also
be slower.
}
\examples{
path <- tempfile("iris", , ".csv")
write.csv(iris, path)
hash_uri(path)

## Note that a different serialization gives a different hash:
path_txt <- tempfile("iris", , ".txt")
write.table(iris, path_txt)
hash_uri(path_txt)

   
}
