% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdb_registry.R
\name{init_lmdb}
\alias{init_lmdb}
\title{inititialize an LMDB registry}
\usage{
init_lmdb(path = default_lmdb())
}
\arguments{
\item{path}{path where LMDB registry should be created}
}
\description{
inititialize an LMDB registry
}
