% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{latest}
\alias{latest}
\title{Access the latest content at a URL}
\usage{
latest(url, dir = content_dir())
}
\arguments{
\item{url}{a URL to a web resource}

\item{dir}{path to the local store directory}
}
\description{
\verb{[latest]} will download the first time it is run, and then use a local cache
unless content has changed. This behavior is similar to \code{pins::pin()},
but uses cryptographic content hashes. Because content hashes are computed in a fast public
content registry, this will usually be faster than downloading on a local connection,
but slower than checking eTags in headers.  Use \link{resolve}
}
\details{
at this time, \code{latest} cannot process FTP resources
}
\examples{
\donttest{

url <- paste0("https://data.giss.nasa.gov/gistemp/graphs/graph_data/",
       "Global_Mean_Estimates_based_on_Land_and_Ocean_Data/graph.txt")
x <- latest(url)

## 
## latest() will always check for changes. Therefore, it is faster to request
## content by resolve() :
resolve(
"hash://sha256/ce1865032089bec8e8a5b0b572c2cb2e3e89bfa8b824520136232d286a908f86",
store = TRUE)
    
}

}
\seealso{
resolve
}
