% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{lookup}
\alias{lookup}
\title{lookup a Content URI or a URL with remote and/or local registries}
\usage{
lookup(uri, registries = c("local", "remote"), ...)
}
\arguments{
\item{uri}{a Content URI or a regular URL for a data file}

\item{registries}{list of registries at which to register the URL}

\item{...}{additional arguments to \verb{[lookup_local]} or \verb{[lookup_remote]}.}
}
\value{
a data frame with matching results
}
\description{
lookup a Content URI or a URL with remote and/or local registries
}
\examples{
\donttest{
  
 ## A content hash
 lookup("hash://sha256/9412325831dab22aeebdd674b6eb53ba6b7bdd04bb99a4dbb21ddff646287e37")
 ## Or a (registered) URL
 lookup("http://cdiac.ornl.gov/ftp/trends/co2/vostok.icecore.co2")
  
  }

}
