% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{lookup_local}
\alias{lookup_local}
\title{look up a Content URI or URL in the local registry}
\usage{
lookup_local(x, dir = app_dir())
}
\arguments{
\item{x}{a URL or the path to a local file}

\item{dir}{the path we should use for permanent / on-disk storage. An appropriate
default will be selected (also configurable using the environmental variable `CONTENTURI_HOME`),
if not specified.}
}
\value{
a data frame with matching results
}
\description{
look up a Content URI or URL in the local registry
}
\examples{
\donttest{
  
 ## A content hash
 lookup_local("hash://sha256/9412325831dab22aeebdd674b6eb53ba6b7bdd04bb99a4dbb21ddff646287e37")
 ## Or a (registered) URL
 lookup_local("http://cdiac.ornl.gov/ftp/trends/co2/vostok.icecore.co2")
  
  }
}
