% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{lookup_remote}
\alias{lookup_remote}
\title{return registered downloadable URLs for the resource from hash-archive.org}
\usage{
lookup_remote(hash)
}
\arguments{
\item{hash}{A hash URI, or download URL of a resource}
}
\value{
a data frame with columns for `url`, `timestamp`, 
`status`, `type` (mimeType), `length` (bytes), and `hashes`
}
\description{
return registered downloadable URLs for the resource from hash-archive.org
}
\examples{
\donttest{
sources <- lookup_remote(
"hash://sha256/9412325831dab22aeebdd674b6eb53ba6b7bdd04bb99a4dbb21ddff646287e37")

# try reading in the first source
df <- read.table(sources$url[1])

# Can also lookup a URL to see if it has been registered:


}
}
