% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_local}
\alias{query_local}
\title{look up a Content URI or URL in the local registry}
\usage{
query_local(uri, dir = app_dir())
}
\arguments{
\item{uri}{a content identifier or a regular URL for a data file}

\item{dir}{the path we should use for permanent / on-disk storage of the registry. An appropriate
default will be selected (also configurable using the environmental variable \code{CONTENTURI_HOME}),
if not specified.}
}
\value{
a data frame with matching results
}
\description{
look up a Content URI or URL in the local registry
}
\examples{
\donttest{
  
 ## A content hash
 query_local("hash://sha256/9412325831dab22aeebdd674b6eb53ba6b7bdd04bb99a4dbb21ddff646287e37")
 ## Or a (registered) URL
 query_local("https://zenodo.org/record/3678928/files/vostok.icecore.co2")
  
  }
}
