% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_remote}
\alias{query_remote}
\title{return registered downloadable URLs for the resource from hash-archive.org}
\usage{
query_remote(uri)
}
\arguments{
\item{uri}{a content identifier or a regular URL for a data file}
}
\value{
a data frame with columns for \code{identifier}, \code{source},
and \code{date}.
}
\description{
return registered downloadable URLs for the resource from hash-archive.org
}
\examples{
\donttest{

query_remote(
"hash://sha256/9412325831dab22aeebdd674b6eb53ba6b7bdd04bb99a4dbb21ddff646287e37")

# Can also query a URL to see if it has been registered:
query_remote("https://zenodo.org/record/3678928/files/vostok.icecore.co2")



}
}
