% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register}
\alias{register}
\title{register a URL with remote and/or local registries}
\usage{
register(url, registries = c("local", "remote"), ...)
}
\arguments{
\item{url}{a URL for a data file}

\item{registries}{list of registries at which to register the URL}

\item{...}{additional arguments to `[local_registry]` or `[remote_registry]`.}
}
\value{
the [httr::response] object for the request (invisibly)
}
\description{
register a URL with remote and/or local registries
}
\examples{
\donttest{
  
 register("http://cdiac.ornl.gov/ftp/trends/co2/vostok.icecore.co2")
  
  }
  
}
