% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_local}
\alias{register_local}
\title{register a URL in a local registry}
\usage{
register_local(url, dir = app_dir())
}
\arguments{
\item{url}{a URL to a remote data resource}

\item{dir}{the path we should use for permanent / on-disk storage of the registry. An appropriate
default will be selected (also configurable using the environmental variable \code{CONTENTURI_HOME}),
if not specified.}
}
\value{
the \link[httr:response]{httr::response} object for the request (invisibly)
}
\description{
register a URL in a local registry
}
\examples{
\donttest{
  
 register_local("https://zenodo.org/record/3678928/files/vostok.icecore.co2")
 
  }
  
}
