% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_remote}
\alias{register_remote}
\title{register a URL with hash-archive.org}
\usage{
register_remote(url)
}
\arguments{
\item{url}{a URL for a data file}
}
\value{
the \link[httr:response]{httr::response} object for the request (invisibly)
}
\description{
register a URL with hash-archive.org
}
\examples{
\donttest{
  
 register_remote("http://cdiac.ornl.gov/ftp/trends/co2/vostok.icecore.co2")
  
  }
  
}
