% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register_url}
\alias{register_url}
\title{register a URL with hash-archive.org}
\usage{
register_url(url)
}
\arguments{
\item{url}{a download URL for a data file}
}
\value{
the [httr::response] object for the request (invisibly)
}
\description{
register a URL with hash-archive.org
}
\examples{
\donttest{
  
  global_mean_temp <- 
  paste0("https://data.giss.nasa.gov/",
         "gistemp/graphs/graph_data/",
         "Global_Mean_Estimates_based_on_Land_and_Ocean_Data/graph.txt")
 
 register_url(global_mean_temp)
  
  }
  
}
