% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{remote_register}
\alias{remote_register}
\title{register a URL with hash-archive.org}
\usage{
remote_register(url)
}
\arguments{
\item{url}{a URL for a data file}
}
\value{
the [httr::response] object for the request (invisibly)
}
\description{
register a URL with hash-archive.org
}
\examples{
\donttest{
  
 remote_register("http://cdiac.ornl.gov/ftp/trends/co2/vostok.icecore.co2")
  
  }
  
}
