% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash-uri.R
\name{resolve_hash}
\alias{resolve_hash}
\title{return registered downloadable URLs for the resource}
\usage{
resolve_hash(hash)
}
\arguments{
\item{hash}{A hash URI, or download URL of a resource}
}
\value{
a data frame with columns for `url`, `timestamp`, 
`status`, `type` (mimeType), `length` (bytes), and `hashes`
}
\description{
return registered downloadable URLs for the resource
}
\examples{
\donttest{
sources <- resolve_hash("hash://sha256/9412325831dab22aeebdd674b6eb53ba6b7bdd04bb99a4dbb21ddff646287e37")

# try reading in the first source
df <- read.table(sources$url[1])
}
}
