% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve an object using the content identifier}
\usage{
retrieve(
  uri,
  prefer = c("local", "url"),
  verify = TRUE,
  registries = c("remote", "local"),
  ...
)
}
\arguments{
\item{uri}{A content identifier or URL}

\item{prefer}{Order of preference if multiple matches are found. See details.}

\item{verify}{logical, default \verb{[TRUE]}. Should we verify the
returned content matches the requested hash?}

\item{registries}{list of registries at which to register the URL}

\item{...}{additional arguments to \verb{[query_local]} or \verb{[query_remote]}.}
}
\description{
Requested content can frequently be found at mutiple locations: cached to disk,
or available at one or more URLs.  This function provides a mechanism to always
return a single, local path to the content requested, (provided the content identifier
can be found in at least one of the registries).
}
\details{
preference
}
\seealso{
query
}
