% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/software-heritage.R
\name{retrieve_swh}
\alias{retrieve_swh}
\title{retrieve content from Software Heritage given a content identifier}
\usage{
retrieve_swh(id, host = "https://archive.softwareheritage.org")
}
\arguments{
\item{id}{a content identifier}

\item{host}{the domain name for the Software Heritage API}
}
\description{
retrieve content from Software Heritage given a content identifier
}
