% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{sources}
\alias{sources}
\title{List all known URL sources for a given Content URI}
\usage{
sources(id, registries = default_registries(), ...)
}
\arguments{
\item{id}{a content identifier}

\item{registries}{list of registries at which to register the URL}

\item{...}{additional arguments}
}
\value{
a data frame with all registration events when a URL or
a local path (including the local store) have contained the corresponding content.
}
\description{
List all known URL sources for a given Content URI
}
\examples{
\donttest{

sources(paste0("hash://sha256/9412325831dab22aeebdd",
             "674b6eb53ba6b7bdd04bb99a4dbb21ddff646287e37"))

}

}
\seealso{
history register store
}
