% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/software-heritage.R
\name{sources_swh}
\alias{sources_swh}
\title{List software heritage sources for a content identifier}
\usage{
sources_swh(id, host = "https://archive.softwareheritage.org", ...)
}
\arguments{
\item{id}{a content identifier}

\item{host}{the domain name for the Software Heritage API}

\item{...}{additional arguments}
}
\description{
List software heritage sources for a content identifier
}
\seealso{
\verb{[sources]}
}
