% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store}
\alias{store}
\title{Store files in a local cache}
\usage{
store(x, dir = content_dir())
}
\arguments{
\item{x}{a URL, \link{connection}, or file path.}

\item{dir}{the path we should use for permanent / on-disk storage
of the registry. An appropriate default will be selected (also
configurable using the environmental variable \code{CONTENTID_HOME}),
if not specified.}
}
\value{
the content-based identifier
}
\description{
Resources at a specified URL will be downloaded and copied
into the local content-based storage. Local paths will simply
be copied into local storage. Identical content is not duplicated.
}
\examples{

# Store & retrieve local file
vostok_co2 <- system.file("extdata", "vostok.icecore.co2",
                          package = "contentid")
 id <- store(vostok_co2)
 retrieve(id)
 
\donttest{
 # Store and retrieve content from a URL 
id <- store("http://cdiac.ornl.gov/ftp/trends/co2/vostok.icecore.co2")
retrieve(id)
}

}
\seealso{
retrieve
}
