% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/software-heritage.R
\name{store_swh}
\alias{store_swh}
\title{Add content to the Software Heritage Archival Store}
\usage{
store_swh(
  origin_url,
  host = "https://archive.softwareheritage.org",
  type = "git",
  ...
)
}
\arguments{
\item{origin_url}{The url address to a GitHub, GitLab, or other recognized repository origin}

\item{host}{the domain name for the Software Heritage API}

\item{type}{software repository type, i.e. "git", "svn"}

\item{...}{additional arguments}
}
\description{
Add content to the Software Heritage Archival Store
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
store_swh("https://github.com/CSSEGISandData/COVID-19")
}
\dontshow{\}) # examplesIf}
}
