% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swh_ratelimit.R
\name{swh_ratelimit}
\alias{swh_ratelimit}
\title{software heritage rate limit}
\usage{
swh_ratelimit(verbose = TRUE)
}
\arguments{
\item{verbose}{show messages about current rate limits?
Software Heritage has a rate limit that can interfere with common queries
(resolve sources, register) when large numbers of queries are made.
contentid functions will not automatically check against wait limits,
but may fall back to other registeries when available.}
}
\value{
remaining, total, and time until next reset are invisibly
returned as a data.frame.
}
\description{
software heritage rate limit
}
