% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{history}
\alias{history}
\title{List all content identifiers that have been seen at a given URL}
\usage{
history(url, registries = default_registries(), ...)
}
\arguments{
\item{url}{A URL for a data file}

\item{registries}{list of registries at which to register the URL}

\item{...}{additional arguments}
}
\value{
a data frame with all content identifiers that have been seen
at a given URL.  If the URL is version-stable, this should be a single
identifier.  Note that if multiple identifiers are listed, older content
may no longer be available, though there is a chance it has been registered
to a different url and can be resolved with \verb{[sources]}.
}
\description{
List all content identifiers that have been seen at a given URL
}
\examples{
\donttest{

history("https://zenodo.org/record/3678928/files/vostok.icecore.co2")

}

}
\seealso{
sources
}
