% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{query a Content URI or a URL with remote and/or local registries}
\usage{
query(uri, registries = default_registries(), ...)
}
\arguments{
\item{uri}{a content identifier or a regular URL for a data file}

\item{registries}{list of registries at which to register the URL}

\item{...}{additional arguments}
}
\value{
a data frame with matching results
}
\description{
DEPRECATED, please use \code{\link[=sources]{sources()}} or \code{\link[=history_url]{history_url()}}
}
