% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{register a URL with remote and/or local registries}
\usage{
register(url, registries = default_registries(), ...)
}
\arguments{
\item{url}{a URL for a data file}

\item{registries}{list of registries at which to register the URL}

\item{...}{additional arguments to \verb{[register_local]} or \verb{[register_remote]}.}
}
\value{
the \link[httr:response]{httr::response} object for the request (invisibly)
}
\description{
register a URL with remote and/or local registries
}
\details{
Local registries can be specified as one or more file paths where local
registries should be created.  Usually a given application will want to register in
only one local registry.  For most use cases, the default registry should be sufficent.
}
\examples{
\donttest{
  
 register("http://cdiac.ornl.gov/ftp/trends/co2/vostok.icecore.co2")
  
  }
  
}
