% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register_local}
\alias{register_local}
\title{register a URL (or local file) in a local registry}
\usage{
register_local(x, store = TRUE, dir = app_dir())
}
\arguments{
\item{x}{a URL or the path to a local file}

\item{store}{Should a local content store for this data also be created?}

\item{dir}{the path we should use for permanent / on-disk storage. An appropriate
default will be selected (also configurable using the environmental variable \code{CONTENTURI_HOME}),
if not specified.}
}
\value{
the \link[httr:response]{httr::response} object for the request (invisibly)
}
\description{
register a URL (or local file) in a local registry
}
\details{
If \code{x} is a URL and \code{store = TRUE}, both the URL and the local
storage path will be registered. If \code{x} is a URL and \code{store = FALSE},
only the URL location will be registered and the downloaded copy of the
object will be deleted after it has been registered.  If \code{x} is a local
path and \code{store = FALSE}, this function will throw an error.
}
\examples{
\donttest{
  
 register_local("http://cdiac.ornl.gov/ftp/trends/co2/vostok.icecore.co2")
 
 vostok_co2 <- system.file("extdata", "vostok.icecore.co2", package = "hashuri")
 register_local(vostok_co2)
  
  }
  
}
