% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve files from the local cache}
\usage{
retrieve(id, dir = content_dir())
}
\arguments{
\item{id}{a \link{content_id}}

\item{dir}{the path we should use for permanent / on-disk storage
of the registry. An appropriate default will be selected (also
configurable using the environmental variable \code{CONTENTURI_HOME}),
if not specified.}
}
\value{
path to a local copy of the file.
}
\description{
Retrieve files from the local cache
}
\examples{

# Store & retrieve local file
vostok_co2 <- system.file("extdata", "vostok.icecore.co2",
                          package = "contentid")
 id <- store(vostok_co2)
 retrieve(id)
 
\donttest{
 # Store and retrieve content from a URL 
id <- store("https://zenodo.org/record/3678928/files/vostok.icecore.co2")
retrieve(id)
}

}
\seealso{
store
}
